feat: 트랜잭션 추상화 설계 적용
